requirements.txt
flask
Flask-MySQLdb
python-dotenv
